test: added tests of StatsController using StatsClient

+++ stats/stats-server/src/main/java/ru/practicum/explorewithme/stats/server/mapper/StatMapper.java
@@ -10,7 +10,7 @@ public class StatMapper {
         return StatDTO.builder()
                 .app(stat.getApp())
                 .uri(stat.getUri())
-                .hits((int) stat.getHits())
+                .hits(stat.getHits())
                 .build();
     }
 }
